chore(wtr): extract shared parts of config into base config object
